pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "aisentinel"
version = "0.1.0"
description = "AISentinel: Data collection, sentiment analysis, and dashboard for AI monitoring."
authors = [{ name = "AISentinel Contributors" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "requests>=2.32",
  "pydantic>=2.9",
  "python-dotenv>=1.0",
  "scikit-learn>=1.5",
  "plotly>=5.24",
  "fastapi>=0.115",
  "uvicorn>=0.30"
]

[project.urls]
Homepage = "https://github.com/your-org/aisentinel"
Repository = "https://github.com/your-org/aisentinel"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "D"
]
ignore = [
  "D203",
  "D213"
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true
